requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
mysqlclient==2.0.3
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.2
